feat: unify mode logic and add metrics
