+++ signals/ai_pattern_filter.py
@@ -9,6 +9,10 @@ import matplotlib.pyplot as plt
 import numpy as np
 
 from ai.cnn_pattern import infer
+from backend.utils import env_loader
+from monitoring import prom_exporter
+
+PROB_THRESHOLD = float(env_loader.get_env("CNN_PROB_THRESHOLD", "0.65"))
 
 
 def _candles_to_image(candles: Iterable[Mapping]) -> np.ndarray:
@@ -35,7 +39,13 @@ def pass_pattern_filter(candles: Iterable[Mapping]) -> tuple[bool, float]:
     img = _candles_to_image(candles)
     res = infer.predict(img)
     prob = res.get("pattern", 0.0)
-    return prob > 0.65, prob
+    passed = prob > PROB_THRESHOLD
+    if passed:
+        try:
+            prom_exporter.increment_pattern_filter_pass()
+        except Exception:
+            pass
+    return passed, prob
 
 
 __all__ = ["pass_pattern_filter"]
